# doc/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mavka Open API Specification",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Dev server"
    }
  ],
...
  "paths": {
...
    "/users/{username}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get user data",
        "description": "This method returns public information about a user",
        "operationId": "getUser",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the user to retrieve information",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "first_name": {
                      "type": "string"
                    },
                    "last_name": {
                      "type": "string"
                    },
                    "wall_id": {
                      "type": "integer"
                    },
                    "avatar_url": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "username": "gervant",
                    "first_name": "Герольд",
                    "last_name": "Ривский",
                    "wall_id": 6,
                    "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "User with username USERNAME does not exist"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change user data",
        "description": "This method returns public information about a user",
        "operationId": "editUser",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "pass": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  }
                },
                "required": ["email", "pass", "first_name", "last_name"]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the user for updating",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "User updated successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "msg": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          },
                          "location": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "message": "Incorrect values error",
                    "errors": [
                      {
                        "type": "field",
                        "msg": "Password can not be empty",
                        "path": "pass",
                        "location": "body"
                      },
                      {
                        "type": "field",
                        "msg": "Password length is incorrect - must be 4 to 64 symbols",
                        "path": "pass",
                        "location": "body"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Can not access private information about another user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Access denied"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email is already used",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "The email is already used"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete user",
        "description": "This method deletes user",
        "operationId": "deleteUser",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the user to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "User deleted"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Can not delete another user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Access denied"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/walls/group-walls": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all group walls",
        "description": "This method returns a list of all group walls",
        "operationId": "getGroupWalls",
        "tags": ["Walls"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "wall_id": {
                        "type": "integer"
                      },
                      "wall_group_id": {
                        "type": "integer"
                      },
                      "wall_img": {
                        "type": "integer"
                      },
                      "wall_title": {
                        "type": "string"
                      },
                      "wall_theme": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "wall_id": 61,
                      "wall_group_id": 7,
                      "wall_img": "https://i.ibb.co/KVwYr15/327893448828.webp",
                      "wall_title": "This is wall title",
                      "wall_theme": "GAMES"
                    },
                    {
                      "wall_id": 62,
                      "wall_group_id": 8,
                      "wall_img": "https://i.ibb.co/KVwYr15/327893448828.webp",
                      "wall_title": "This is wall title",
                      "wall_theme": "GAMES"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/walls/search-walls": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all walls with the theme and title starting with title",
        "description": "This method returns a list group walls filtered by themes and title",
        "operationId": "searchGroupWalls",
        "tags": ["Walls"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  }
                },
                "required": ["theme", "title"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "wall_id": {
                        "type": "integer"
                      },
                      "wall_group_id": {
                        "type": "integer"
                      },
                      "wall_img": {
                        "type": "integer"
                      },
                      "wall_title": {
                        "type": "string"
                      },
                      "wall_theme": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "wall_id": 61,
                      "wall_group_id": 7,
                      "wall_img": "https://i.ibb.co/KVwYr15/327893448828.webp",
                      "wall_title": "This is wall title",
                      "wall_theme": "GAMES"
                    },
                    {
                      "wall_id": 62,
                      "wall_group_id": 8,
                      "wall_img": "https://i.ibb.co/KVwYr15/327893448828.webp",
                      "wall_title": "This is wall title",
                      "wall_theme": "GAMES"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/walls/wall-themes": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get wall thiemes",
        "description": "This method returns a list of all wall themes",
        "operationId": "getWallThemes",
        "tags": ["Walls"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "themes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": {
                  "themes": [
                    "Юмор",
                    "Спорт",
                    "Игры",
                    "Еда",
                    "Музыка",
                    "Фотография",
                    "Финансы",
                    "Информационные технологии",
                    "Кино",
                    "Наука"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}